clean: move store js files to "admin" folder
