+++ js/app/admin/store/list.js
@@ -1,4 +1,4 @@
-import i18n from '../i18n'
+import i18n from '../../i18n'
 
 document.querySelectorAll('.delete-store').forEach((el) => {
     el.addEventListener('click', (e) => {
